test: BoardTest에서 GameTest를 분리
